+++ lib/client/isoexpress/mra_capacity_results.dart
@@ -1,5 +1,60 @@
+import 'dart:convert';
+
 import 'package:date/date.dart';
+import 'package:dotenv/dotenv.dart' as dotenv;
+import 'package:elec/elec.dart';
 import 'package:timezone/timezone.dart';
+import 'package:http/http.dart' as http;
+
+Future<List<MraCapacityZoneRecord>> getMraClearingPriceZone(Month month) async {
+  final url = '${dotenv.env['RUST_SERVER']}/isone/capacity/mra/results/zone'
+      '/start/${month.toIso8601String()}/end/${month.toIso8601String()}';
+  final response = await http.get(Uri.parse(url));
+  if (response.statusCode != 200) {
+    throw Exception(
+        'Failed to load MRA clearing price: ${response.statusCode}');
+  }
+  final List<dynamic> jsonList = jsonDecode(response.body);
+  return jsonList.map((e) {
+    final month = Month.fromInt(e['month'] as int, location: IsoNewEngland.location);
+    return MraCapacityZoneRecord(
+        month: month,
+        capacityZoneId: e['capacity_zone_id'] as int,
+        capacityZoneType: e['capacity_zone_type'] as String,
+        capacityZoneName: e['capacity_zone_name'] as String,
+        supplyOffersSubmitted: e['supply_offers_submitted'] as num,
+        demandBidsSubmitted: e['demand_bids_submitted'] as num,
+        supplyOffersCleared: e['supply_offers_cleared'] as num,
+        demandBidsCleared: e['demand_bids_cleared'] as num,
+        netCapacityCleared: e['net_capacity_cleared'] as num,
+        clearingPrice: e['clearing_price'] as num);
+  }).toList();
+}
+
+Future<List<MraCapacityInterfaceRecord>> getMraClearingPriceInterface(Month month) async {
+  final url = '${dotenv.env['RUST_SERVER']}/isone/capacity/mra/results/interface'
+      '/start/${month.toIso8601String()}/end/${month.toIso8601String()}';
+  final response = await http.get(Uri.parse(url));
+  if (response.statusCode != 200) {
+    throw Exception(
+        'Failed to load MRA interface clearing prices for ${month.toIso8601String()}: ${response.statusCode}');
+  }
+  final List<dynamic> jsonList = jsonDecode(response.body);
+  return jsonList.map((e) {
+    final month = Month.fromInt(e['month'] as int, location: IsoNewEngland.location);
+    return MraCapacityInterfaceRecord(
+        month: month,
+        externalInterfaceName: e['external_interface_name'] as String,
+        externalInterfaceId: e['external_interface_id'] as int,
+        supplyOffersSubmitted: e['supply_offers_submitted'] as num,
+        demandBidsSubmitted: e['demand_bids_submitted'] as num,
+        supplyOffersCleared: e['supply_offers_cleared'] as num,
+        demandBidsCleared: e['demand_bids_cleared'] as num,
+        netCapacityCleared: e['net_capacity_cleared'] as num,
+        clearingPrice: e['clearing_price'] as num);
+  }).toList();
+}
+
 
 enum ResourceType {
   generating,