+++ test/db/isoexpress/mra_capacity_bidoffer_test.dart
@@ -1,4 +1,5 @@
 import 'package:dotenv/dotenv.dart' as dotenv;
+import 'package:elec_server/client/isoexpress/mra_capacity_bidoffer.dart';
 import 'package:elec_server/src/db/lib_prod_archives.dart';
 import 'package:elec_server/src/db/lib_prod_dbs.dart';
 import 'package:test/test.dart';
@@ -11,6 +12,17 @@ Future<void> tests(String rootUrl) async {
     test('read file for 2024-01', () async {
       var file = archive.getFilename(Month.utc(2024, 1));
       var data = archive.processJsonFile(file);
+      expect(data.length, 454);
+      var xs = data.where((e) => e.maskedResourceId == 52995).toList();
+      expect(xs.length, 5);
+      var segments = xs.map((e) => e.segment).toList();
+      segments.sort();
+      expect(segments, [0, 1, 2, 3, 4]);
+    });
+  });
+  group('MRA BidOffer client tests:', () {
+    test('get data for 2024-01', () async {
+      var data = await getMraBidsOffers(Month.utc(2024, 1));
       expect(data.length, 454);
       var xs = data.where((e) => e.maskedResourceId == 52995).toList();
       expect(xs.length, 5);