+++ CHANGELOG.md
@@ -7,6 +7,9 @@
 - Move ISONE Monthly Asset NCPC to DuckDB
 - test
 
+# 2026-07-02
+- Add client functions for ISONE MRA results and bids/offers.
+
 # 2026-06-24
 - Add HQ fuel mix client